+++ app/src/main/java/com/example/complamap/views/activities/MainActivity.kt
@@ -16,7 +16,7 @@ import android.view.inputmethod.InputMethodManager
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.core.view.isGone
 import androidx.core.view.isVisible
-import com.example.complamap.ListFragment
+import com.example.complamap.views.fragments.ListFragment
 import com.example.complamap.R
 import com.example.complamap.views.fragments.MapFragment
 import com.example.complamap.views.fragments.PhotoFragment
