+++ app/src/main/java/com/example/complamap/views/fragments/ListFragment.kt
@@ -1,4 +1,4 @@
-package com.example.complamap
+package com.example.complamap.views.fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
